consolidated quiz
